Rename webpack.js to webpack.base.config.js
